Add details component and some details
