+++ src/App.js
@@ -3,7 +3,7 @@ import title from './craftybeer-title.png'
 import './App.css';
 import Search from './components/Search'
 import { Routes, Route, Link } from 'react-router-dom'
-import BrewDetails from './BrewDetails';
+import BrewDetails from './components/BrewDetails';
 import Header from './components/Header'
 
 
@@ -13,7 +13,6 @@ const App =() => {
     <>
       <Header />
       <main className='main-container'>
-        {/* <Search />       */}
       <Routes>
         <Route path='/' element={<Search />} />
         <Route path='/:brewyId' element={<BrewDetails />} />
