+++ src/components/css/BrewDetails.css
@@ -1,22 +1,54 @@
 * {
   margin: 0;
+
 }
 
-.brewyDetailsContainer {
+.temp {
+  display: flex;
+  flex-direction: column;
+  align-items:center;
+  width: 50%;
+}
+
+.brewy-details {
   display: flex;
   flex-direction: column;
   align-items: center;
-  width:50%;
-  height: 70%;
+  width:70%;
+  height: 35%;
   color: white;
- 
 }
 
 .name {
   font-size: 50px;
   margin:10px;
+  color: #dbb26f;
 }
 
 article p {
+  font-size: 40px;
+  padding:10px;
+}
+
+a {
   font-size: 30px;
+  color: #f2c963;
+
+}
+
+button {
+  width: 10em;
+  height: 3em;
+  font-size: 20px;
+  border:none;
+  background-color: #dbb26f;
+  border-radius: 10px;
+}
+
+button:hover {
+  background-color: #af3f26;
+}
+
+.links-section {
+  display: flex;
 }
